Fix: Mismatching frame ids in the header & poses of exploration path
The code works only if the Global Frame id of the costmap is given as "map" and if any other global frame id is provided, there is a mismatch in the frame id string in the "header of path topic" and the "header of poses in the path topic". This creates error if we try to visualize the path in rviz window.

Hence I have modified the code to directly obtain the frame id from costmap_2d_ros_ object to fix the mismatch. Now the code works perfectly for me, for any given global frame id.

And thank you so much for providing this wonderful package as open source.